Add WiFi connectivity support for Raspberry Pi Pico W with separate module
